Movie(Enum)
this my first times using enum and it really is interesting how it works. It took me a while to understand what it really does but I just overcomplicated it. Overall though this program assignment does use something that could involve a real world situation of people who go the movies to see some movie at a certain time. I was brainstorming on ideas to create a program specifically to work with things like people who pay for movies, when to come in, best prices, best times, ect... But it would take away from my time from learning more about python so you know I will just hold off on it right now but again this was fun to do.